# api/openapi-bundled.yaml
openapi: 3.1.0
info:
  title: Focus API
  description: Personal task management (projects, tasks, tags, reviews) with Google sign-in and an MCP server.
  version: 1.0.0
  contact:
    name: Cosmic Teacups
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://focus-api.cosmicteacups.com
    description: Production server
security:
  - cookieAuth: []
...
paths:
...
  /api/dates/parse:
    post:
      tags:
        - tasks
      summary: Parse a natural-language date
      description: Resolves a phrase to a timestamp using the rule-based parser, falling back to Claude for anything it cannot handle.
      operationId: parseDate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParseDateRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParseDateResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - cookieAuth: []
  /api/repeats/parse:
    post:
      tags:
        - tasks
      summary: Parse a natural-language repeat schedule
      description: Resolves a phrase to a repeat rule using the rule-based parser, falling back to Claude, without changing anything.
      operationId: parseRepeat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParseRepeatRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParseRepeatResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - cookieAuth: []
...
components:
  schemas:
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Error message
          example: An error occurred
...
    RepeatUnit:
      type: string
      enum:
        - day
        - week
        - month
        - year
    RepeatRule:
      type: object
      description: How a task repeats. "from" controls whether the next occurrence is scheduled relative to the completion date or the previous due/defer date. Optional anchors pin occurrences to calendar days.
      required:
        - every
        - unit
        - from
      properties:
        every:
          type: integer
          minimum: 1
          example: 1
        unit:
          $ref: '#/components/schemas/RepeatUnit'
        from:
          type: string
          enum:
            - completion
            - due
          example: completion
        weekdays:
          type: array
          description: Only with unit=week — which weekdays (0 = Sunday … 6 = Saturday)
          items:
            type: integer
            minimum: 0
            maximum: 6
        day_of_month:
          type: integer
          description: Only with unit=month — day of the month (1–31) or -1 for the last day; days past a month's end clamp to its last day
          minimum: -1
          maximum: 31
...
    ParseRepeatRequest:
      type: object
      required:
        - input
      properties:
        input:
          type: string
          example: first of every month
        timezone:
          type: string
          example: America/Los_Angeles
        reference:
          type: string
          format: date-time
          description: The reference "today"; defaults to now
    ParseRepeatResponse:
      type: object
      required:
        - rule
        - description
        - next_occurrences
        - source
      properties:
        rule:
          $ref: '#/components/schemas/RepeatRule'
        description:
          type: string
          example: monthly on the 1st
        first_occurrence:
          type: string
          format: date-time
          description: For calendar-anchored rules, the first matching date on or after today
        next_occurrences:
          type: array
          description: The next few occurrences, starting from first_occurrence (or today)
          items:
            type: string
            format: date-time
        source:
          type: string
          enum:
            - rules
            - ai
...
    ParseDateRequest:
      type: object
      required:
        - input
      properties:
        input:
          type: string
          example: next friday
        kind:
          type: string
          enum:
            - defer
            - due
          description: Picks the default time of day when the phrase has none (defer = start of day, due = 17:00)
        timezone:
          type: string
          example: America/New_York
        reference:
          type: string
          format: date-time
          description: The reference "now" for relative phrases; defaults to the current time
    ParseDateResponse:
      type: object
      required:
        - resolved_at
        - interpretation
        - source
      properties:
        resolved_at:
          type: string
          format: date-time
        interpretation:
          type: string
          description: Human-readable explanation of how the phrase was interpreted
          example: Monday, October 5 at 12:00 AM
        source:
          type: string
          enum:
            - rules
            - ai
          description: Whether the deterministic parser or the AI fallback resolved the phrase
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_id
      description: Session ID cookie set after Google sign-in